Added: pages component and react router
