+++ package.json
@@ -10,6 +10,7 @@
     "hamburger-react": "^2.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
