+++ src/components/pages/About.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const About = () => {
+  return (
+    <div>
+      <h1>Info about JTL</h1>
+    </div>
+  );
+};
