+++ src/components/pages/Home.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+export const Home = () => {
+  return (
+    <div>
+      <div className="container">
+        <h1>Welcome to Jane's Trading</h1>
+      </div>
+    </div>
+  );
+};
