+++ src/components/pages/Products.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Products = () => {
+  return (
+    <div>
+      <h1>Products page</h1>
+    </div>
+  );
+};
