+++ src/components/pages/index.js
@@ -1,0 +1,3 @@
+export * from "./Home";
+export * from "./About";
+export * from "./Products";
